requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
SQLAlchemy
marshmallow
marshmallow-sqlalchemy
flask-marshmallow
psycopg2